Creating an anonymous feedback request
